viki-ai/entity_extraction/pyproject.toml
[project]
name = "entity-extraction"
version = "0.1.0"
description = "Entity extraction service using FastAPI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.0",
    "google-cloud-firestore>=2.19.0",
    "google-cloud-logging>=3.8.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-tasks>=2.16.0",
    "google-cloud-monitoring-dashboards>=1.12.0",
    "google-genai[aiohttp]>=0.8.0",
    "coloredlogs>=15.0.1",
    "aiocache>=0.12.2",
    "pypdf>=4.0.0",
    "httpx>=0.24.0",
    "opentelemetry-api>=1.21.0,<1.29.0",
    "opentelemetry-sdk>=1.21.0,<1.29.0",
    "opentelemetry-semantic-conventions>=0.43b0,<0.50b0",
    "opentelemetry-instrumentation-fastapi>=0.43b0,<0.50b0",
    "opentelemetry-instrumentation-httpx>=0.43b0,<0.50b0",
    "opentelemetry-instrumentation-requests>=0.43b0,<0.50b0",
    "opentelemetry-exporter-gcp-trace>=1.6.0",
    "opentelemetry-exporter-jaeger>=1.21.0,<1.22.0",
    "opentelemetry-exporter-otlp>=1.21.0,<1.29.0",
    "pyjwt[crypto]>=2.8.0",
    "setuptools<81",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "httpx>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
pythonpath = ["src"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
